convert routing to use query string params instead - MSAL limitation
