+++ components/WorkOrderSystem.vue
@@ -212,7 +212,7 @@
         <!-- <v-icon size="small" class="me-2" @click="editItem(item.raw)">
           mdi-pencil
         </v-icon> -->
-        <NuxtLink :to="{ path: '/workorders/' + item.raw.id }" target="_blank">Edit work order</NuxtLink>
+        <NuxtLink :to="'/workorders?id=' + item.raw.id" target="_blank">Edit work order</NuxtLink>
       </template>
 
       <template v-slot:bottom v-if="!showFooter"></template>
